requirements.txt
groq
keyboard
pyaudio
pyperclip
pystray
pillow
python-dotenv
plyer
requests
pyautogui
